requirements.txt
requests
python-dotenv
aiohttp
watchdog
pillow
regex
huggingface_hub
aiosqlite
discord.py==2.5.2
flask==3.1.1